Move the redux stuff to JS side
